+++ frontend/src/api/authentication.ts
@@ -1,9 +1,7 @@
+import { apiClient, customFetch } from '@/api';
 import { END_POINTS } from '@/routes';
 import type { LoginRequest, SignupRequest } from '@/types';
 import { MemberInfo } from '@/types';
-
-import { apiClient } from './ApiClient';
-import { customFetch } from './customFetch';
 
 const API_URL = process.env.REACT_APP_API_URL || 'https://default-url.com';
