+++ frontend/src/api/index.ts
@@ -1,7 +1,7 @@
-export { CATEGORY_API_URL, getCategoryList, postCategory, deleteCategory } from './categories';
-export { TAG_API_URL, getTagList } from './tags';
-export { customFetch } from './customFetch';
-export { QUERY_KEY } from './queryKeys';
+export { TAG_API_URL, getTagList } from '@/api/tags';
+export { customFetch } from '@/api/customFetch';
+export { QUERY_KEY } from '@/api/queryKeys';
+export { CATEGORY_API_URL, getCategoryList, postCategory, editCategory, deleteCategory } from '@/api/categories';
 export {
   TEMPLATE_API_URL,
   PAGE_SIZE,
@@ -13,7 +13,7 @@ export {
   postTemplate,
   editTemplate,
   deleteTemplate,
-} from './templates';
+} from '@/api/templates';
 export {
   CHECK_NAME_API_URL,
   LOGIN_API_URL,
@@ -25,6 +25,9 @@ export {
   postLogout,
   getLoginState,
   checkName,
-} from './authentication';
-export { LIKE_API_URL, postLike, deleteLike } from './like';
-export { getMemberName } from './members';
+} from '@/api/authentication';
+export { LIKE_API_URL, postLike, deleteLike } from '@/api/like';
+export { getMemberName } from '@/api/members';
+export { ApiError } from '@/api/Error/ApiError';
+export { HTTP_STATUS } from '@/api/Error/statusCode';
+export { apiClient } from '@/api/ApiClient';
