+++ frontend/src/api/like.ts
@@ -1,9 +1,8 @@
 import { HttpResponse } from 'msw';
 
+import { customFetch } from '@/api/customFetch';
 import { END_POINTS } from '@/routes';
 import { LikeDeleteRequest, LikePostRequest } from '@/types';
-
-import { customFetch } from './customFetch';
 
 const API_URL = process.env.REACT_APP_API_URL || 'https://default-url.com';
